# feat!: variant internal stepper data