… Core/include/Acts/Propagator/StraightLineStepper.hpp
… Core/include/Acts/Propagator/detail/SteppingHelper.hpp
… Core/include/Acts/TrackFitting/KalmanFitter.hpp
… Core/src/Propagator/StraightLineStepper.cpp
… Core/src/Propagator/detail/CovarianceEngine.cpp
… Tests/UnitTests/Core/Propagator/AtlasStepperTests.cpp
… Tests/UnitTests/Core/Propagator/CMakeLists.txt
… Tests/UnitTests/Core/Propagator/CovarianceEngineTests.cpp
… Tests/UnitTests/Core/Propagator/EigenStepperTests.cpp
… Tests/UnitTests/Core/Propagator/NavigatorTests.cpp
… Tests/UnitTests/Core/Propagator/StraightLineStepperTests.cpp