This PR supersedes #727 and

 * prepares the internal data types of the Steppers to do all parameterisations
 * renames the transportJacobian methods consistently to what they do