@asalzburger 2021-03-09
first timing measurement:

* master (w/o) this first development

```shell
time ./bin/ActsBenchmarkEigenStepper
16:13:20    Eigen_Steppe   INFO      propagating 20000 tracks with pT = 1GeV in a 2T B-field
16:13:24    Eigen_Steppe   INFO      Execution stats: 20000 runs of 1 iteration(s), 1520.5ms total, 75.8210+/-0.0010µs per run, 75821.000+/-1.043ns per iteration
16:13:24    Eigen_Steppe   INFO      average path length = 5000mm
./bin/ActsBenchmarkEigenStepper  3.52s user 0.00s system 99% cpu 3.525 total
```
* with variant internal types

```shell
time ./bin/ActsBenchmarkEigenStepper
15:36:59    Eigen_Steppe   INFO      propagating 20000 tracks with pT = 1GeV in a 2T B-field
15:37:03    Eigen_Steppe   INFO      Execution stats: 20000 runs of 1 iteration(s), 1483.4ms total, 73.6310+/-0.0057µs per run, 73631.000+/-5.666ns per iteration
15:37:03    Eigen_Steppe   INFO      average path length = 5000mm
./bin/ActsBenchmarkEigenStepper  3.48s user 0.00s system 99% cpu 3.488 total
```